Split the workbench home spec to smaller once.
